requirements.txt
# Core LLM & RAG
langchain==0.3.7
langchain-openai==0.2.8
langchain-community==0.3.7
openai==1.54.0

# Vector store
faiss-cpu==1.9.0
chromadb==0.5.18

# Data
pandas==2.2.3
numpy==1.26.4
faker==30.8.2

# App
streamlit==1.40.0

# Utils
python-dotenv==1.0.1
tiktoken==0.8.0

# Notebook
jupyterlab==4.3.1
ipykernel==6.29.5sentence-transformers 
langchain-huggingface 
langchain-groq 
